Calculate tempo durations

+++ MidiParser/include/MidiParser.h
@@ -54,7 +54,7 @@ private:
     inline T ReadInteger(T* destination = nullptr);
     inline void ReadBytes(void* buffer, size_t size);
 
-    inline uint64_t TicksToMicroseconds(uint32_t ticks, uint32_t tempo);
+    inline float TicksToMicroseconds(uint32_t ticks, uint32_t tempo);
 
     inline uint32_t CalculateTempo(uint8_t* data, size_t size);
